# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Kuwaiti Tailoring Shop SaaS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shop/suborders:
    post:
      operationId: createSubOrder
      tags: [suborders]
      summary: Create a sub-order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubOrderBody"
      responses:
        "201":
          description: Sub-order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubOrder"
  /shop/suborders/{subOrderId}:
    patch:
      operationId: updateSubOrder
      tags: [suborders]
      summary: Update sub-order
      parameters:
        - name: subOrderId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubOrderBody"
      responses:
        "200":
          description: Sub-order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubOrder"
...
  /shop/users:
    get:
      operationId: listShopUsers
      tags: [users]
      summary: List shop users (manager only)
      responses:
        "200":
          description: Shop users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopUserListResponse"
    post:
      operationId: createShopUser
      tags: [users]
      summary: Create shop user (manager only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateShopUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopUser"
  /shop/users/{userId}:
    patch:
      operationId: updateShopUser
      tags: [users]
      summary: Update shop user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateShopUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopUser"
    delete:
      operationId: deleteShopUser
      tags: [users]
      summary: Delete shop user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Measurements:
      type: object
      properties:
        id:
          type: integer
        profileId:
          type: integer
        length:
          type: ["number", "null"]
        shoulder:
          type: ["number", "null"]
        chest:
          type: ["number", "null"]
        sleeve:
          type: ["number", "null"]
        neck:
          type: ["number", "null"]
        modelNotes:
          type: ["string", "null"]
        generalNotes:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - profileId
        - updatedAt
...
    SubOrder:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        shopId:
          type: integer
        profileId:
          type: integer
        profileName:
          type: string
        subOrderNumber:
          type: string
        quantity:
          type: integer
        fabricSource:
          type: string
          enum: [shop_fabric, customer_fabric]
        fabricDescription:
          type: ["string", "null"]
        price:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [under_tailoring, ready]
        createdAt:
          type: string
          format: date-time
        measurements:
          $ref: "#/components/schemas/Measurements"
      required:
        - id
        - invoiceId
        - shopId
        - profileId
        - profileName
        - subOrderNumber
        - quantity
        - fabricSource
        - price
        - paidAmount
        - remainingAmount
        - status
        - createdAt
...
    CreateSubOrderBody:
      type: object
      properties:
        invoiceId:
          type: integer
        profileId:
          type: integer
        quantity:
          type: integer
        fabricSource:
          type: string
          enum: [shop_fabric, customer_fabric]
        fabricDescription:
          type: ["string", "null"]
        price:
          type: number
        paidAmount:
          type: number
        notes:
          type: ["string", "null"]
      required:
        - invoiceId
        - profileId
        - quantity
        - fabricSource
        - price
        - paidAmount
...
    UpdateSubOrderBody:
      type: object
      properties:
        quantity:
          type: ["integer", "null"]
        fabricSource:
          type: ["string", "null"]
          enum: [shop_fabric, customer_fabric, null]
        fabricDescription:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        paidAmount:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
...
    ShopUser:
      type: object
      properties:
        id:
          type: integer
        shopId:
          type: integer
        username:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [shop_manager, reception, tailor]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - shopId
        - username
        - name
        - role
        - createdAt
    ShopUserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/ShopUser"
      required:
        - users
    CreateShopUserBody:
      type: object
      properties:
        username:
          type: string
        name:
          type: string
        password:
          type: string
        role:
          type: string
          enum: [shop_manager, reception, tailor]
      required:
        - username
        - name
        - password
        - role
    UpdateShopUserBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        password:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
          enum: [shop_manager, reception, tailor, null]
...